+++ packages/learn-card-types/aqu.config.ts
@@ -1,0 +1,3 @@
+const options = { declaration: 'normal' };
+
+export default options;